+++ build_executable.py
@@ -10,7 +10,7 @@ import subprocess
 import sys
 from pathlib import Path
 
-from pyhabitat.version_info import get_python_version, form_dynamic_binary_name
+from pyhabitat.packaging import get_python_version, form_dynamic_binary_name
 from pyhabitat._version import get_version
 from pyhabitat.system_info import SystemInfo
 from pyhabitat.environment import on_windows